+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed a bug where element editors were showing provisional changes, even if the user didn’t have permission to save them.
 - Fixed a bug where the control panel could be inaccessible if a mutex lock couldn’t be acquired for the queue. ([#13052](https://github.com/craftcms/cms/issues/13052))
 - Fixed a bug where it wasn’t possible to update a Matrix block on a revision without a new block ID being assigned. ([#13064](https://github.com/craftcms/cms/discussions/13064))
+- Fixed a JavaScript error that could occur on field layout designers, if any tabs didn’t have any elements. ([#13062](https://github.com/craftcms/cms/issues/13062))
 - Fixed XSS vulnerabilities.
 - Updated yii2-debug to 2.1.22. ([#13058](https://github.com/craftcms/cms/issues/13058))